+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 Change log for NuMPI
 ===================
 
-v0.10.0 (not yet released)
---------------------------
+v0.10.0 (24Apr26)
+-----------------
 
-- ENH: L-BFGS with linear constraint
+- ENH: L-BFGS with simple box bounds
+- ENH: L-BFGS with linear constraint (and box bounds)
 - ENH: CCG now supports gradient-only objective callables via `jac=False`
   and split objective/gradient callables via `jac=<callable>`
+- BUILD: Switched build system to hatchling and hatch-vcs
 
 v0.9.0 (24Nov25)
 ----------------